+++ src/signing/index.ts
@@ -15,3 +15,4 @@ export * from "./primitives";
 export * from "./universalSign";
 export * from "./types";
 export * from "./codexStrategy";
+export * from "./partialSig";